Add save! method on LuckyRecord::Form
This method will work like the normal save, but only take parameters,
without a block. If it saves successfully, it will return the record,
_not_ the form. If it fails, it will raise a
LuckyRecord::Exceptions::InvalidForm error.

I tried to use the existing logic where it made sense, since the goal is
for the method to work the same but have some different behavior at the
end of the save.