Add Queryable#empty?
I find that I'm looking for this functionality when I want to show
something on a page based on the presence of a record. For example, an
empty-state UI that shows "No posts yet" as opposed to listing out all
posts. This could check a PostQuery.new.empty? to determine whether to
loop through posts or show a sad face.

I also thought about checking for `first?` but that returns a record if
it's present. That could be worked around with `!...nil?` but at that
point `size.zero?` seemed like a pretty solid choice. Also, maybe it's
faster?